+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.14.0] - 2023-08-14
+
+### Changed
+
+- cmsdev
+  - Simplified `lib.common.Restful()` function
+  - Added good path BOS API GET tests with tenant specified for supported v2 endpoints
+  - Updated v2 sessions CLI, session templates CLI (v1 and v2), and v1 session templates API tests to
+    handle multi-tenancy in their responses from BOS (while not including it in their queries to BOS).
+
 ## [1.13.0] - 2023-08-10
 
 ### Changed
@@ -208,7 +218,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-[Unreleased]: https://github.com/Cray-HPE/cms-tools/compare/1.13.0...HEAD
+[Unreleased]: https://github.com/Cray-HPE/cms-tools/compare/1.14.0...HEAD
+
+[1.14.0]: https://github.com/Cray-HPE/cms-tools/compare/1.13.0...1.14.0
 
 [1.13.0]: https://github.com/Cray-HPE/cms-tools/compare/1.12.0...1.13.0